fixing incorrect types and starting branch instrructions

+++ include/instructions/s.h
@@ -14,7 +14,7 @@
 
 class S{
     private:
-        using func = std::function<void(uint16_t, uint8_t, uint16_t)>; //<returnType(params)>
+        using func = std::function<void(int32_t, uint8_t, uint8_t)>; //<returnType(params)>
         std::unordered_map<uint8_t, func> sInstructions;
         Registers reg;
         std::array<uint32_t, 32>& registers;
@@ -23,9 +23,9 @@ class S{
     public:
         S();
         void findInstruction(uint8_t, uint8_t, uint8_t, uint8_t, uint8_t);
-        void execute_sb(uint16_t, uint8_t, uint8_t);
-        void execute_sh(uint16_t, uint8_t, uint8_t);
-        void execute_sw(uint16_t, uint8_t, uint8_t);
+        void execute_sb(int32_t, uint8_t, uint8_t);
+        void execute_sh(int32_t, uint8_t, uint8_t);
+        void execute_sw(int32_t, uint8_t, uint8_t);
 };
 
 #endif
